Add Olimexpc board

+++ src/emupriv.h
@@ -7,11 +7,11 @@ typedef struct
   uint64_t total_time;
   uint32_t underrun;
   int32_t  sound_count;
-  int32_t  int_count;
+  int64_t  int_count;
 } perform;
 
-extern uint32_t sound_count;
-extern uint32_t int_count;
+extern int32_t sound_count;
+extern int64_t int_count;
 
 #endif
 #endif
